requirements.txt
python-telegram-bot>=20.7
openai>=1.12.0
PyPDF2>=3.0.0
python-docx>=1.1.0
python-dotenv>=1.0.0

